Redirect the root page to the "articles/"
